fix: 沙箱缺陷批量修复 - IPC超时生效/IsAvailable去死锁/net10.0/降级链/网络隔离警告/Linux检测/结构体布局/git只读

+++ core/safety/Guard/src/Security/Sandbox/Native/JobObjectNative.cs
@@ -48,7 +48,7 @@ internal static class JobObjectNative
         internal nint MinimumWorkingSetSize;
         internal nint MaximumWorkingSetSize;
         internal uint ActiveProcessLimit;
-        internal long Affinity;
+        internal nint Affinity;
         internal uint PriorityClass;
         internal uint SchedulingClass;
     }
